+++ HISTORY.md
@@ -2,6 +2,7 @@ unreleased
 ==========
 
   * Add `priority` option
+  * Fix `expires` option to reject invalid dates
 
 0.4.2 / 2022-02-02
 ==================